requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
Pillow>=8.3.0
tqdm>=4.62.0
matplotlib>=3.5.0
seaborn>=0.11.0